webservice_deployment/requirements.txt
pydantic
fastapi
mlflow==2.4
psutil==5.9.5
scikit-learn==1.2.2
uvicorn
google-cloud-storage
python-dotenv
httpx